+++ CHANGES.md
@@ -1,4 +1,4 @@
-## [0.5.1] current
+## [0.6.0] current
 
 ### Added
 
@@ -10,6 +10,7 @@
 ### Changes
 
 - Removed `Module.unload`, instead `Module.load_data_ex` attaches an unload as a finalizer.
+- Removed `Deviceptr.mem_free`, instead it is attached as a finalizer.
 
 ## [0.5.0] 2024-09-25